Move compute_samples to netket.sampler and update tests
