+++ Sources/Utils/py_utils.cc
@@ -14,6 +14,9 @@
 
 #include "py_utils.hpp"
 
+#include <pybind11/eigen.h>
+
+#include "Sampler/vmc_sampling.hpp"
 #include "Utils/all_utils.hpp"
 
 namespace py = pybind11;
@@ -37,6 +40,8 @@ void AddUtilsModule(py::module m) {
       },
       py::arg("seed") = netket::default_random_engine::default_seed,
       R"EOF(seed: The chosen seed for the distributed random number generator.  )EOF");
+
+  subm.def("_subtract_mean", &SubtractMean, py::arg{"values"}.noconvert());
 
   py::class_<MPIHelpers>(m, "MPI")
       .def_static("rank", &MPIHelpers::MPIRank,
